# Final optimization batch (#15): use memory pools